+++ uniffi_bindgen/src/bindings/kotlin/templates/RecordTemplate.kt
@@ -7,7 +7,7 @@ data class {{ type_name }} (
     {%- call kt::docstring(field, 4) %}
     {% if config.generate_immutable_records() %}val{% else %}var{% endif %} {{ field.name()|var_name }}: {{ field|type_name(ci) -}}
     {%- match field.default_value() %}
-        {%- when Some(literal) %} = {{ literal|render_literal(field, ci) }}
+        {%- when Some(default) %} = {{ default|render_default(field, ci) }}
         {%- else %}
     {%- endmatch -%}
     {% if !loop.last %}, {% endif %}